+++ Oracle.Services/Interfaces/IProveedorService.cs
@@ -10,9 +10,9 @@ namespace Oracle.Services.Interfaces
     public interface IProveedorService
     {
         Task<RespuestaService<List<Proveedor>>> Listar();
-        Task<RespuestaService<Proveedor>> BuscarPorId(decimal id);
+        Task<RespuestaService<Proveedor>> BuscarPorId(int id);
         Task<RespuestaService<Proveedor>> Guardar(Proveedor pr);
         Task<RespuestaService<Proveedor>> Actualizar(Proveedor pr);
-        Task<RespuestaService<bool>> Eliminar(decimal id);
+        Task<RespuestaService<bool>> Eliminar(int id);
     }
 }
